+++ rosetta-source/src/test/java/cdm/observable/asset/fro/functions/IndexValueObservationMultipleTest.java
@@ -1,9 +1,8 @@
 package cdm.observable.asset.fro.functions;
 
-import cdm.observable.asset.FloatingRateIndex;
 import cdm.observable.asset.InterestRateIndex;
-import cdm.observable.asset.functions.IndexValueObservation;
-import cdm.observable.asset.functions.IndexValueObservationMultiple;
+import cdm.observable.asset.calculatedrate.functions.IndexValueObservation;
+import cdm.observable.asset.calculatedrate.functions.IndexValueObservationMultiple;
 import com.google.inject.Binder;
 import com.google.inject.Inject;
 import com.rosetta.model.lib.records.Date;
