+++ rosetta-source/src/test/java/cdm/product/asset/floatingrate/functions/DetermineFloatingRateResetTest.java
@@ -2,7 +2,7 @@ package cdm.product.asset.floatingrate.functions;
 
 import cdm.base.datetime.BusinessCenterEnum;
 import cdm.observable.asset.InterestRateIndex;
-import cdm.observable.asset.functions.IndexValueObservation;
+import cdm.observable.asset.calculatedrate.functions.IndexValueObservation;
 import cdm.product.asset.InterestRatePayout;
 import cdm.product.asset.RateSpecification;
 import cdm.product.asset.calculation.functions.GetNotionalAmountTest;
