+++ rosetta-source/src/test/java/cdm/product/asset/floatingrate/functions/FloatingRateTestHelper.java
@@ -6,8 +6,8 @@ import cdm.base.staticdata.asset.rates.FloatingRateIndexEnum;
 import cdm.observable.asset.FloatingRateIndex;
 import cdm.observable.asset.InterestRateIndex;
 import cdm.observable.asset.PriceSchedule;
+import cdm.observable.asset.calculatedrate.functions.IndexValueObservation;
 import cdm.observable.asset.fro.functions.IndexValueObservationTestDataProvider;
-import cdm.observable.asset.functions.IndexValueObservation;
 import cdm.product.asset.FloatingRateSpecification;
 import cdm.product.asset.SpreadSchedule;
 import cdm.product.common.schedule.RateSchedule;
